Dev/algopix api

+++ database/migrations/2020_12_28_212813_create_supplier_algopix_table.php
@@ -53,7 +53,7 @@ class CreateSupplierAlgopixTable extends Migration
             $table->float('msrp_amount')->nullable();
             $table->enum('msrp_currency',['USD'])->nullable();
         
-            $table->tinyInteger('demand')->nullable();
+            $table->smallInteger('demand')->nullable();
 
             //o means offers 
             $table->string('o_market_brand')->nullable()->comment('o for Offers');
@@ -78,27 +78,27 @@ class CreateSupplierAlgopixTable extends Migration
             $table->string('o_msd_object_type')->nullable()->comment('o for Offers, msd for Market Specific Data');
             $table->enum('o_msd_estimated_sales_revenues_currency',['USD'])->nullable()->comment('o for Offers, msd for Market Specific Data');
             $table->float('o_msd_estimated_sales_revenues_amount')->nullable()->comment('o for Offers, msd for Market Specific Data');
-            $table->tinyInteger('o_msd_estimated_unit_sold')->nullable()->comment('o for Offers, msd for Market Specific Data');
+            $table->integer('o_msd_estimated_unit_sold')->nullable()->comment('o for Offers, msd for Market Specific Data');
 
             //c means competition
             $table->enum('o_msd_c_level', ['HIGH', 'MEDIUM', 'LOW'])->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
-            $table->tinyInteger('o_msd_c_number_of_offers')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
+            $table->integer('o_msd_c_number_of_offers')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
         
             //lowest_ofrs means lowest Offer From Reputable Seller
             $table->enum('o_msd_c_lofrs_price_currency_code', ['USD'])->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, lofrs for Lowest Offer From Reputable Seller');
             $table->float('o_msd_c_lofrs_price_amount')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, lofrs for Lowest Offer From Reputable Seller');
             $table->string('o_msd_c_lofrs_srating_positive_feedback_rating')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, lofrs for Lowest Offer From Reputable Seller, srating for Seller Rating');
-            $table->tinyInteger('o_msd_c_lofrs_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, lofrs for Lowest Offer From Reputable Seller, srating for Seller Rating');
+            $table->integer('o_msd_c_lofrs_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, lofrs for Lowest Offer From Reputable Seller, srating for Seller Rating');
         
             $table->enum('o_msd_c_lowest_offer_price_currency_code', ['USD'])->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
             $table->float('o_msd_c_lowest_offer_price_amount')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
             $table->string('o_msd_c_lowest_offer_srating_positive_feedback_rating')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, srating for Seller Rating');
-            $table->tinyInteger('o_msd_c_lowest_offer_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, srating for Seller Rating');
+            $table->integer('o_msd_c_lowest_offer_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, srating for Seller Rating');
 
             $table->enum('o_msd_c_buy_box_offering_price_currency_code', ['USD'])->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
             $table->float('o_msd_c_buy_box_offering_price_amount')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
             $table->string('o_msd_c_buy_box_offering_srating_positive_feedback_rating')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition, srating for Seller Rating');
-            $table->tinyInteger('o_msd_c_buy_box_offering_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
+            $table->integer('o_msd_c_buy_box_offering_srating_number_of_seller_ratings')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
 
             $table->string('o_msd_c_amazon_offer_state')->nullable()->comment('o for Offers, msd for Market Specific Data, c for Competition');
             $table->string('o_msd_local_asin')->nullable()->comment('o for Offers, msd for Market Specific Data');
@@ -107,7 +107,7 @@ class CreateSupplierAlgopixTable extends Migration
             $table->string('o_msd_amazon_category_id')->nullable()->comment('o for Offers, msd for Market Specific Data');
             $table->integer('o_msd_amazon_category_best_seller_ranking')->nullable()->comment('o for Offers, msd for Market Specific Data');
             $table->string('o_msd_solid_via_fba')->nullable()->comment('o for Offers, msd for Market Specific Data');
-            $table->tinyInteger('o_msd_number_off_offers_sold_via_fba')->nullable()->comment('o for Offers, msd for Market Specific Data');
+            $table->integer('o_msd_number_off_offers_sold_via_fba')->nullable()->comment('o for Offers, msd for Market Specific Data');
 
             $table->string('o_image_url')->nullable()->comment('o for Offers');
